Move summary functions out of client.py
